removed uncatched exception

+++ aikif/programs.py
@@ -111,10 +111,8 @@ class Programs(object):
         #location_fileList = filemap.get_full_filename(filemap.find_type('LOCATION'), filemap.find_ontology('FILE-PROGRAM')[0])     
         object_fileList = filemap.get_full_filename(filemap.find_type('OBJECT'), filemap.find_ontology('FILE-PROGRAM')[0])      
         print('object_fileList = ' + object_fileList + '\n')
-        try:
+        if os.path.exists(object_fileList):
             os.remove(object_fileList)
-        except Exception:
-            pass
         self.lstPrograms.sort()
 
         try:
